+++ src/main/java/fll/web/GatherBugReport.java
@@ -20,6 +20,7 @@ import java.sql.SQLException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Set;
+import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 
@@ -33,6 +34,7 @@ import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
 
 import fll.Utilities;
+import fll.Version;
 import fll.db.DumpDB;
 import fll.xml.ChallengeDescription;
 
@@ -83,6 +85,12 @@ public class GatherBugReport extends BaseFLLServlet {
                                    application.getMajorVersion(), application.getMinorVersion(), //
                                    application.getServerInfo())
                            .getBytes(Utilities.DEFAULT_CHARSET));
+
+        zipOut.putNextEntry(new ZipEntry("fll-sw_version.txt"));
+        final String versionInfo = Version.getAllVersionInformation().entrySet().stream() //
+                                          .map(e -> String.format("%s:%s", e.getKey(), e.getValue())) //
+                                          .collect(Collectors.joining("\n"));
+        zipOut.write(versionInfo.getBytes(Utilities.DEFAULT_CHARSET));
 
         addDatabase(zipOut, connection, challengeDescription);
         addLogFiles(zipOut);