+++ Changes.md
@@ -1,4 +1,24 @@
-Release 15.6.0
+Release 15.9.1
+==============
+* #911, #918 - Allow the awards report to be output when there are no performance scores
+
+Release 15.9.0
+==============
+* #912 - Use the order of subjective categories in the challenge description for the awards report
+* #914 - Fix parsing of LocalTime in javascript by using the JSJoda time package
+* #905 - Make tab to skip over read-only cells on non-numeric categories page
+* #904 - Add FLL-SW version information to bug reports
+* #902 - Fix loading of non-numeric award winners into the edit award winners page
+* #903 - Fix importing of finalist schedule information when loading a database
+* #903 - properly handle a tournament without a date
+
+Release 15.8.0
+==============
+* #900 - Focus on the text field after opening the comment dialog in the subjective application
+* #899 - Make sure that all properties exist in the subjective judging application all of the time
+* Fix bug where the SQL query page doesn't work
+
+Release 15.7.0
 ==============
 * #897 - Fix handling of repeating question marks in comments
 * #892 - Color comment buttons in the subjective application when comments have been entered